+++ src/ir/ExpressionAnalyzer.cpp
@@ -192,7 +192,6 @@ bool ExpressionAnalyzer::flexibleEqual(Expression* left,
 #define DELEGATE_FIELD_INT(id, field) COMPARE_FIELD(field)
 #define DELEGATE_FIELD_LITERAL(id, field) COMPARE_FIELD(field)
 #define DELEGATE_FIELD_NAME(id, field) COMPARE_FIELD(field)
-#define DELEGATE_FIELD_SIGNATURE(id, field) COMPARE_FIELD(field)
 #define DELEGATE_FIELD_TYPE(id, field) COMPARE_FIELD(field)
 #define DELEGATE_FIELD_HEAPTYPE(id, field) COMPARE_FIELD(field)
 #define DELEGATE_FIELD_ADDRESS(id, field) COMPARE_FIELD(field)
@@ -314,7 +313,6 @@ struct Hasher {
 
 #define DELEGATE_FIELD_INT(id, field) HASH_FIELD(field)
 #define DELEGATE_FIELD_LITERAL(id, field) HASH_FIELD(field)
-#define DELEGATE_FIELD_SIGNATURE(id, field) HASH_FIELD(field)
 
 #define DELEGATE_FIELD_NAME(id, field) visitNonScopeName(cast->field)
 #define DELEGATE_FIELD_TYPE(id, field) visitType(cast->field);